Allow services to specify an IP address as host.
Make up an internal hostname UUID.<s.domain>. which is used in the
SRV records. Include an A record in the additional section which
lists the IP address originally specified.

This way the need for an external nameserver is diminished.

Resolver might not grok the additional section, but I think a seperate
SkyDNS client API is needed anyway.