Add no-expire flag to not auto evict records on ttl
There is a scaling issue with ttl and heartbeats.  In the case of
docker you can have 1000s of containers running and with a lower ttl in
a dymanic environment it can become very chatty updating ttls for all
the records.  In the case of docker and skydock we know when a service
is removed so there is no need for the overhead of heartbeats because
docker will take care of that for us.  Also there is a race trying to
update 1000s of records that the heartbeat does not hit the cluster
before the ttl is automatically expired.